+++ packages/orq-rc/RELEASES.md
@@ -10649,3 +10649,13 @@ Based on:
 - [python v4.15.0-rc.17] packages/orq-rc
 ### Releases
 - [PyPI v4.15.0-rc.17] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.17 - packages/orq-rc
+
+## 2026-09-03 11:15:22
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.3 (2.934.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.15.0-rc.18] packages/orq-rc
+### Releases
+- [PyPI v4.15.0-rc.18] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.18 - packages/orq-rc